Diff: changed type of gridsize to size_t.

+++ src/util.h
@@ -194,7 +194,7 @@ int     cdoDefineZaxis(const char *zaxisfile);
 
 int     vlistInqNWPV(int vlistID, int varID);
 int     vlistIsSzipped(int vlistID);
-int     vlist_check_gridsize(int vlistID);
+size_t  vlist_check_gridsize(int vlistID);
 int     vlist_get_psvarid(int vlistID, int zaxisID);
 double *vlist_read_vct(int vlistID, int *rzaxisIDh, int *rnvct, int *rnhlev, int *rnhlevf, int *rnhlevh);
 void vlist_change_hybrid_zaxis(int vlistID1, int vlistID2, int zaxisID1, int zaxisID2);
